refactor: dedupe cover and poster mappers
