+++ projects/client/src/lib/requests/_internal/mapMovieResponseToMovieSummary.ts
@@ -1,16 +1,9 @@
 import type { MovieCertificationResponse, MovieResponse } from '$lib/api.ts';
-import { mediumUrl } from '$lib/requests/_internal/mediumUrl.ts';
-import { thumbUrl } from '$lib/requests/_internal/thumbUrl.ts';
 import type { MovieSummary } from '$lib/requests/models/MovieSummary.ts';
-import {
-  DEFAULT_TRAILER,
-  MAX_DATE,
-  MEDIA_COVER_LARGE_PLACEHOLDER,
-  MEDIA_COVER_THUMB_PLACEHOLDER,
-  MEDIA_POSTER_PLACEHOLDER,
-} from '$lib/utils/constants.ts';
-import { findDefined } from '$lib/utils/string/findDefined.ts';
+import { DEFAULT_TRAILER, MAX_DATE } from '$lib/utils/constants.ts';
 import { prependHttps } from '$lib/utils/url/prependHttps.ts';
+import { mapCover } from './mapCover';
+import { mapPoster } from './mapPoster';
 
 function mapMovieCertificationResponse(
   certification?: MovieCertificationResponse,
@@ -26,15 +19,8 @@ function mapMovieCertificationResponse(
 export function mapMovieResponseToMovieSummary(
   movie: MovieResponse,
 ): MovieSummary {
-  const posterCandidate = findDefined(
-    movie.images?.poster.at(1),
-    movie.images?.poster.at(0),
-  );
-
-  const coverCandidate = findDefined(
-    movie.images?.fanart.at(1),
-    movie.images?.fanart.at(0),
-  );
+  const poster = mapPoster(movie.images);
+  const cover = mapCover(movie.images);
 
   return {
     id: movie.ids.trakt,
@@ -45,35 +31,10 @@ export function mapMovieResponseToMovieSummary(
     tagline: movie.tagline ?? '',
     country: movie.country,
     languages: movie.languages,
-    poster: {
-      url: {
-        medium: prependHttps(
-          mediumUrl(posterCandidate),
-          MEDIA_POSTER_PLACEHOLDER,
-        ),
-        thumb: prependHttps(
-          thumbUrl(posterCandidate),
-          MEDIA_POSTER_PLACEHOLDER,
-        ),
-      },
-    },
-    cover: {
-      url: {
-        medium: prependHttps(
-          mediumUrl(coverCandidate),
-          MEDIA_COVER_LARGE_PLACEHOLDER,
-        ),
-        thumb: prependHttps(
-          thumbUrl(coverCandidate),
-          MEDIA_COVER_THUMB_PLACEHOLDER,
-        ),
-      },
-    },
+    poster,
+    cover,
     thumb: {
-      url: prependHttps(
-        thumbUrl(coverCandidate),
-        MEDIA_COVER_THUMB_PLACEHOLDER,
-      ),
+      url: cover.url.thumb,
     },
     genres: movie.genres ?? [],
     status: movie.status ?? 'unknown',
